Made tunnel timeout configurable, increased default to 30 minutes.  Previous 250ms resulted in premature socket close, trancated files.
git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/branches/branch_1-68@44180 4f837ed2-42f5-46e7-a7a5-fa17313484d4

+++ src/org/lockss/proxy/ProxyHandler.java
@@ -290,10 +290,10 @@ public class ProxyHandler extends AbstractHttpHandler {
    * IE on win2000 has connections issues with normal timeout handling.
    * This timeout should be set to a low value that will expire to allow IE to
    * see the end of the tunnel connection.
-   * /
-   public void setTunnelTimeoutMs(int ms) {
-   _tunnelTimeoutMs = ms;
-   }
+   */
+  public void setTunnelTimeoutMs(long ms) {
+    _tunnelTimeoutMs = (int)ms;
+  }
 
    /* ------------------------------------------------------------ */
   public void handle(String pathInContext,
